cleaned up and separated styling properties
